requirements.txt
flask==1.1.2
flask-Session
flask-login
flask-sqlalchemy
flask-socketio
flask-cors
eventlet
pytest
redis
rq